+++ drivers/docker/testing_unix.go
@@ -1,0 +1,61 @@
+// +build darwin dragonfly freebsd linux netbsd openbsd solaris
+
+package docker
+
+import (
+	"fmt"
+	"io"
+	"os"
+	"path/filepath"
+
+	"github.com/hashicorp/nomad/client/allocdir"
+	testing "github.com/mitchellh/go-testing-interface"
+)
+
+func newTaskConfig(variant string, command []string) TaskConfig {
+	// busyboxImageID is the ID stored in busybox.tar
+	busyboxImageID := "busybox:1.29.3"
+
+	image := busyboxImageID
+	loadImage := "busybox.tar"
+	if variant != "" {
+		image = fmt.Sprintf("%s-%s", busyboxImageID, variant)
+		loadImage = fmt.Sprintf("busybox_%s.tar", variant)
+	}
+
+	return TaskConfig{
+		Image:     image,
+		LoadImage: loadImage,
+		Command:   command[0],
+		Args:      command[1:],
+	}
+}
+
+func copyImage(t testing.T, taskDir *allocdir.TaskDir, image string) {
+	dst := filepath.Join(taskDir.LocalDir, image)
+	copyFile(filepath.Join("./test-resources/docker", image), dst, t)
+}
+
+// copyFile moves an existing file to the destination
+func copyFile(src, dst string, t testing.T) {
+	in, err := os.Open(src)
+	if err != nil {
+		t.Fatalf("copying %v -> %v failed: %v", src, dst, err)
+	}
+	defer in.Close()
+	out, err := os.Create(dst)
+	if err != nil {
+		t.Fatalf("copying %v -> %v failed: %v", src, dst, err)
+	}
+	defer func() {
+		if err := out.Close(); err != nil {
+			t.Fatalf("copying %v -> %v failed: %v", src, dst, err)
+		}
+	}()
+	if _, err = io.Copy(out, in); err != nil {
+		t.Fatalf("copying %v -> %v failed: %v", src, dst, err)
+	}
+	if err := out.Sync(); err != nil {
+		t.Fatalf("copying %v -> %v failed: %v", src, dst, err)
+	}
+}
